requirements.txt
Django==1.7
Markdown==2.6.2
PyYAML==3.11
Werkzeug==0.10.4
chardet==2.2.1
colorama==0.2.5
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.0
django-extensions==1.5.2
django-jsonview==0.4.3
django-markdown==0.8.4
html5lib==0.999
nose==1.3.1
psycopg2==2.5.3
requests==2.2.1
selenium==2.45.0
six==1.5.2
sqlparse==0.1.15
uWSGI==2.0.10
urllib3==1.7.1
wheel==0.24.0
